Removed unused code, moved show-tree to api. Resolves #9 and #158
