--- front-end-includes/notifications.php
@@ -1,7 +1,0 @@
-<div class="notification-wrapper">
-  <a href="<?php echo $currentPage;?>/results.php#results" title="Go to results">Go to download section</a>
-  <button title="Close this message">
-    <i class="fa fa-times fa-fw fa-loaded" aria-hidden="true"></i>
-    <span class="sr-only">Close this notification</span>
-  </button>
-</div>
