--- front-end-includes/results-table-wrapper.php
@@ -1,17 +1,0 @@
-<div class="table-wrapper">
-  <table>
-    <thead>
-      <tr>
-          <th>#</th>
-          <th>ID</th>
-          <th>Component</th>
-          <th>Sentence</th>
-      </tr>
-    </thead>
-    <tbody>
-    </tbody>
-    <tbody class="empty">
-        <tr><td colspan="4">No results found yet</td></tr>
-    </tbody>
-  </table>
-</div>
