--- front-end-includes/secondary-navigation.php
@@ -1,3 +1,0 @@
-<nav class="secondary-navigation">
-    <?php buildProgressList(); ?>
-</nav>
