--- front-end-includes/tb-cgn.php
@@ -1,300 +1,0 @@
-<div class="table-wrapper">
-<table>
-    <thead>
-        <tr>
-            <th colspan=2>Treebank</th>
-            <th>Contents</th>
-            <th>Sentences</th>
-            <th>Words</th>
-            <th>Sentences</th>
-            <th>Words</th>
-            <th>Sentences</th>
-            <th>Words</th>
-        </tr>
-        <tr>
-            <th>
-                <label>
-                    <input type="checkbox" data-check="all-cgn-nl" checked> NL</label>
-            </th>
-            <th>
-                <label>
-                    <input type="checkbox" data-check="all-cgn-vl" checked> VL</label>
-            </th>
-            <th></th>
-            <th colspan=2>NL</th>
-            <th colspan=2>VL</th>
-            <th colspan=2>Total</th>
-        </tr>
-    </thead>
-    <tbody>
-        <tr>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="na" checked> NA</label>
-            </td>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="va" checked> VA</label>
-            </td>
-            <td>Spontaneous conversations ('face-to-face')</td>
-            <td>50,239</td>
-            <td>302,828</td>
-            <td>22,881</td>
-            <td>147,418</td>
-            <td>73,120</td>
-            <td>450,246</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="nb" checked> NB</label>
-            </td>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="vb" checked> VB</label>
-            </td>
-            <td>Interviews with teachers of Dutch</td>
-            <td>2,484</td>
-            <td>25,724</td>
-            <td>4,289</td>
-            <td>34,158</td>
-            <td>6,773</td>
-            <td>59,882</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="nc" checked> NC</label>
-            </td>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="vc" checked> VC</label>
-            </td>
-            <td>Telephone conversations (recorded via a switchboard)</td>
-            <td>11,649</td>
-            <td>70,084</td>
-            <td>3,142</td>
-            <td>19,984</td>
-            <td>14,791</td>
-            <td>90,068</td>
-        </tr>
-        <tr>
-            <td>
-                <label class="disabled">
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="nd" disabled> ND</label>
-            </td>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="vd" checked> VD</label>
-            </td>
-            <td>Telephone conversations (recorded on MD)</td>
-            <td>0</td>
-            <td>0</td>
-            <td>929</td>
-            <td>6,309</td>
-            <td>929</td>
-            <td>6,309</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="ne" checked> NE
-                    <label>
-            </td>
-            <td>
-                <label class="disabled">
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="ve" disabled> VE</label>
-            </td>
-            <td>Simulated business negotiations</td>
-            <td>3,123</td>
-            <td>25,524</td>
-            <td>0</td>
-            <td>0</td>
-            <td>3,123</td>
-            <td>25,524</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="nf" checked> NF</label>
-            </td>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="vf" checked> VF</label>
-            </td>
-            <td>Interviews/discussions/debates (broadcast)</td>
-            <td>6,290</td>
-            <td>75,167</td>
-            <td>2,617</td>
-            <td>25,122</td>
-            <td>8,907</td>
-            <td>100,289</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="ng" checked> NG</label>
-            </td>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="vg" checked> VG</label>
-            </td>
-            <td>(Political) discussions/debates/meetings (non-broadcast)</td>
-            <td>1,166</td>
-            <td>25,125</td>
-            <td>543</td>
-            <td>9,009</td>
-            <td>1,709</td>
-            <td>34,134</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="nh" checked> NH</label>
-            </td>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="vh" checked> VH</label>
-            </td>
-            <td>Lessons recorded in the classroom</td>
-            <td>3,064</td>
-            <td>26,004</td>
-            <td>1,395</td>
-            <td>10,116</td>
-            <td>4,459</td>
-            <td>36,120</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="ni" checked> NI</label>
-            </td>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="vi" checked> VI</label>
-            </td>
-            <td>Live (sports) commentaries (broadcast)</td>
-            <td>2,251</td>
-            <td>25,002</td>
-            <td>1,026</td>
-            <td>10,147</td>
-            <td>3,277</td>
-            <td>35,149</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="nj" checked> NJ</label>
-            </td>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="vj" checked> VJ</label>
-            </td>
-            <td>Newsreports (broadcast)</td>
-            <td>2,259</td>
-            <td>25,084</td>
-            <td>536</td>
-            <td>7,686</td>
-            <td>2,795</td>
-            <td>32,770</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="nk" checked> NK</label>
-            </td>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="vk" checked> VK</label>
-            </td>
-            <td>News (broadcast)</td>
-            <td>1,923</td>
-            <td>25,353</td>
-            <td>558</td>
-            <td>7,306</td>
-            <td>2,481</td>
-            <td>32,659</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="nl" checked> NL</label>
-            </td>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="vl" checked> VL</label>
-            </td>
-            <td>Commentaries/columns/reviews (broadcast)</td>
-            <td>1,857</td>
-            <td>25,082</td>
-            <td>601</td>
-            <td>7,431</td>
-            <td>2,458</td>
-            <td>32,513</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="nm" checked> NM</label>
-            </td>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="vm" checked> VM</label>
-            </td>
-            <td>Ceremonious speeches/sermons</td>
-            <td>444</td>
-            <td>5,190</td>
-            <td>107</td>
-            <td>1,894</td>
-            <td>551</td>
-            <td>7,084</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="nn" checked> NN</label>
-            </td>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="vn" checked> VN</label>
-            </td>
-            <td>Lectures/seminars</td>
-            <td>593</td>
-            <td>14,921</td>
-            <td>701</td>
-            <td>8,159</td>
-            <td>1,294</td>
-            <td>23,080</td>
-        </tr>
-        <tr>
-            <td>
-                <label class="disabled">
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="no" disabled> NO</label>
-            </td>
-            <td>
-                <label>
-                    <input type="checkbox" data-treebank="cgn" name="subtreebank[]" value="vo" checked> VO</label>
-            </td>
-            <td>Read speech</td>
-            <td>0</td>
-            <td>0</td>
-            <td>3,256</td>
-            <td>44,144</td>
-            <td>3,256</td>
-            <td>44,144</td>
-        </tr>
-        <tfoot>
-            <tr>
-                <td colspan=2>CGN treebank</td>
-                <td>Complete treebank</td>
-                <td>87,342</td>
-                <td>671,088</td>
-                <td>42,581</td>
-                <td>338,883</td>
-                <td>129,923</td>
-                <td>1,009,971</td>
-            </tr>
-        </tfoot>
-    </tbody>
-</table>
-</div>
