--- front-end-includes/tb-sonar.php
@@ -1,220 +1,0 @@
-<div class="table-wrapper">
-<table>
-    <thead>
-        <tr>
-            <th>Treebank</th>
-            <th>Contents</th>
-            <th>Sentences</th>
-        </tr>
-    </thead>
-    <tbody>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPEA"> WR-P-E-A</label>
-            </td>
-            <td>Discussion lists</td>
-            <td>4,395,094</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPEC"> WR-P-E-C</label>
-            </td>
-            <td>E-magazines</td>
-            <td>551,119</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPEE"> WR-P-E-E</label>
-            </td>
-            <td>Newsletters</td>
-            <td>115</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPEF"> WR-P-E-F</label>
-            </td>
-            <td>Press releases</td>
-            <td>18,372</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPEG"> WR-P-E-G</label>
-            </td>
-            <td>Subtitles</td>
-            <td>3,925,824</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPEH"> WR-P-E-H</label>
-            </td>
-            <td>Teletext pages</td>
-            <td>40,715</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPEI"> WR-P-E-I</label>
-            </td>
-            <td>Websites</td>
-            <td>204,998</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPEJ" checked> WR-P-E-J</label>
-            </td>
-            <td>Wikipedia</td>
-            <td>1,354,245</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPEK"> WR-P-E-K</label>
-            </td>
-            <td>Blogs</td>
-            <td>8,614</td>
-        </tr>
-        <tr>
-            <td>
-                <label class="disabled">
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPEL" disabled> WR-P-E-L</label>
-            </td>
-            <td>Tweets</td>
-            <td>2,636,859</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPPB"> WR-P-P-B</label>
-            </td>
-            <td>Books</td>
-            <td>1,709,808</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPPC"> WR-P-P-C</label>
-            </td>
-            <td>Brochures</td>
-            <td>74,897</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPPD"> WR-P-P-D</label>
-            </td>
-            <td>Newsletters</td>
-            <td>2,184</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPPE"> WR-P-P-E</label>
-            </td>
-            <td>Guides &amp; manuals</td>
-            <td>19,077</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPPF"> WR-P-P-F</label>
-            </td>
-            <td>Legal texts</td>
-            <td>650,211</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPPG"> WR-P-P-G</label>
-            </td>
-            <td>Newspapers</td>
-            <td>14,973,209</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPPH"> WR-P-P-H</label>
-            </td>
-            <td>Periodicals &amp; magazines</td>
-            <td>5,475,556</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPPI"> WR-P-P-I</label>
-            </td>
-            <td>Policy documents</td>
-            <td>387,312</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPPJ"> WR-P-P-J</label>
-            </td>
-            <td>Proceedings</td>
-            <td>16,935</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRPPK"> WR-P-P-K</label>
-            </td>
-            <td>Reports</td>
-            <td>93,507</td>
-        </tr>
-        <tr>
-            <td>
-                <label class="disabled">
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRUEA" disabled> WR-U-E-A</label>
-            </td>
-            <td>Chats</td>
-            <td>2,387,131</td>
-        </tr>
-        <tr>
-            <td>
-                <label class="disabled">
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRUED" disabled> WR-U-E-D</label>
-            </td>
-            <td>SMS</td>
-            <td>101,116</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WRUEE"> WR-U-E-E</label>
-            </td>
-            <td>Written assignments</td>
-            <td>23,488</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WSUEA"> WS-U-E-A</label>
-            </td>
-            <td>Auto cues</td>
-            <td>2,163,521</td>
-        </tr>
-        <tr>
-            <td>
-                <label>
-                    <input type="radio" data-treebank="sonar" name="subtreebank[]" value="WSUTB"> WS-U-T-B</label>
-            </td>
-            <td>Texts for the visually impaired</td>
-            <td>44,862</td>
-        </tr>
-        <tfoot>
-            <tr>
-                <td>SoNaR treebank</td>
-                <td>Complete treebank</td>
-                <td>41,258,769</td>
-            </tr>
-        </tfoot>
-        <tbody>
-</table>
-</div>
