+++ web-ui/src/app/services/configuration.service.ts
@@ -10,8 +10,13 @@ export class ConfigurationService {
         this.config = this.loadConfig();
     }
 
-    async getApiUrl(path: string): Promise<string> {
-        return (await this.config).apiUrl + path;
+    async getApiUrl(path: string, parts: string[] = [], queryString: { [key: string]: string } = {}): Promise<string> {
+        const queryStringEntries = Object.entries(queryString);
+        return (await this.config).apiUrl + path +
+            (parts.length ? '/' + parts.join('/') : '') +
+            (queryStringEntries.length
+                ? queryStringEntries.map(([key, value], index) => `${index ? '&' : '?'}${key}=${value}`)
+                : '');
     }
 
     async getGretelUrl(path: string): Promise<string> {
@@ -22,13 +27,13 @@ export class ConfigurationService {
         return (await this.config).uploadUrl + path;
     }
 
-    async loadConfig() {
+    private async loadConfig() {
         return this.httpClient.get<Config>(`assets/config/config.${environment.name}.json`).toPromise();
     }
 }
 
 interface Config {
-    "apiUrl": string
-    "gretelUrl": string
-    "uploadUrl": string
+    'apiUrl': string;
+    'gretelUrl': string;
+    'uploadUrl': string;
 }
